Added type, tests and method. Remove unused example project.

+++ NetTelebot/Type/MessageInfo.cs
@@ -41,10 +41,8 @@ namespace NetTelebot.Type
                 ? new UserInfo(jsonObject["from"].Value<JObject>())
                 : new UserInfo();
 
-            DateUnix = jsonObject["date"].Value<int>();
-
+            DateUnix = jsonObject["date"].Value<long>();
             Date = DateUnix.ToDateTime();
-
             Chat = new ChatInfo(jsonObject["chat"].Value<JObject>());
 
             ForwardFrom = jsonObject["forward_from"] != null
@@ -60,7 +58,7 @@ namespace NetTelebot.Type
 
             if (jsonObject["forward_date"]!=null)
             {
-                ForwardDateUnix = jsonObject["forward_date"].Value<int>();
+                ForwardDateUnix = jsonObject["forward_date"].Value<long>();
                 ForwardDate = ForwardDateUnix.ToDateTime();
             }
 
@@ -70,7 +68,7 @@ namespace NetTelebot.Type
 
             if (jsonObject["edit_date"] != null)
             {
-                EditDateUnix = jsonObject["edit_date"].Value<int>();
+                EditDateUnix = jsonObject["edit_date"].Value<long>();
                 EditDate = EditDateUnix.ToDateTime();
             }
